zipit-base-files: switch to a more openwrt-friendly init system and change preinit overlay
The updated init script system uses the original openwrt style minus procd support. Any
initscript that is designed to work with procd will not be enabled or run and instead
displays a message stating such.

The updated preinit runs fsck to check for errors on the SD card before mounting. If
errors are found it will reboot. If booting from SD Card (or USB), preinit mounts the
device as / and continues to regular init (skipping overlay mount).

When booting from internal flash, preinit attempts to mount the jffs. If that fails it
mounts a tmpfs to /overly instead and continues. If /etc/overlay.conf is found in jffs,
it attempts to mount the configured overlay. The configured overlay then becomes the
'upper' filesystem and takes all writes. The jffs is mounted in between root (squashfs)
and the configured overlay so that changes in jffs are applied if they exist. The
configured overlay device root is available in /mnt. All changes are written to the
'upperdir' configured in overlay.conf (see that file for instructions).

Signed-off-by: Joe Honold <mozzwald@gmail.com>