netsurf-all: bump version to 3.6, add patch to allow custom user agent string, and enable Javascript via Duktape.
You can now set the user agent string in /usr/share/netsurf/Choices which
has 3 examples of mobile user agent strings. To use the default netsurf
user agent, remove or comment out the user_agent option. Tested with google
and ebay which both provide the mobile site which looks better on the zipit.

Javascript is now enabled with the builtin Duktape library. eBay login was
possible so it is partially working. Gmail allowed login but complained that
javascript was not enabled.

Signed-off-by: Joe Honold <mozzwald@gmail.com>